add ThreeDotsMenü and transport setFavorites logic to ThreeDotsMenü.js
