+++ components/ArticleCard/styles.js
@@ -12,6 +12,7 @@ export const ContentContainer = styled.div`
   flex-direction: row;
   align-items: flex-start;
   margin-bottom: 15px;
+
 `;
 
 export const StyledH2 = styled.h2`
@@ -28,7 +29,7 @@ export const StyledImageContainer = styled.div`
   border-radius: 30px;
   overflow: hidden;
   margin-bottom: 15px;
-  width: 256; 
+  width: 256;
   height: 144;
 `;
 
@@ -37,34 +38,7 @@ export const ButtonContainer = styled.div`
   flex-direction: column;
   align-items: flex-start;
   margin-left: 10px;
-  gap: 20px;
+  width: 0;
+  height: 0;
 `;
 
-export const StyledLink = styled.a`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 95px;
-  height: 30px;
-  border-radius: 40px;
-  background: rgba(0, 128, 255, 0.9);
-  color: white;
-  cursor: pointer;
-  text-decoration: none;
-  font-size: 13px;
-`;
-
-export const StyledButton = styled.button`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 130px;
-  height: 30px;
-  border-radius: 40px;
-  background: rgba(0, 128, 255, 0.9);
-  color: white;
-  cursor: pointer;
-  text-decoration: none;
-  font-size: 13px;
-  border: none;
-`;
