+++ components/FavoriteButton/styles.js
@@ -4,4 +4,6 @@ export const StyledMarkButton = styled.button`
   background: none;
   border: none;
   cursor: pointer;
+  padding:0px;
+  margin: 0px;
 `;
