+++ components/ThreeDotsMenü/styles.js
@@ -1,0 +1,47 @@
+import styled from "styled-components";
+
+export const StyledLink = styled.a`
+  background: transparent;
+  color: white;
+  cursor: pointer;
+  text-decoration: none;
+  font-size: 13px;
+  border: none;
+`;
+
+export const StyledUL = styled.ul`
+  position: relative;
+  top: -22px;
+  right: 258px;
+  display: flex;
+  width: 200px;
+  height: 91px;
+  padding: 16px 24px;
+  flex-direction: column;
+  align-items: left;
+  gap: 8px;
+  border-radius: 30px;
+  background-blend-mode: luminosity;
+  background: rgba(245, 245, 245, 0.1);
+  backdrop-filter: blur(10px);
+`;
+
+export const StyledLI = styled.li`
+  position: relative;
+  top: -3px;
+  left: 100px;
+  margin: 4px;
+  align-text: right;
+  background: transparent;
+  color: white;
+  cursor: pointer;
+  text-decoration: none;
+  font-size: 13px;
+  border: none;
+`;
+
+export const StyledButton = styled.button`
+  border: none;
+  background: transparent;
+  cursor: pointer;
+`;
