+++ pages/styles.css
@@ -1,7 +1,8 @@
 li {
-    list-style-type: none;
+  list-style-type: none;
 }
 
 a {
-    text-decoration: none;
+  text-decoration: none;
 }
+
